Introduce pyproject-nix overlay in nix flake (#4862)
In the past, when we were using poetry2nix, we used additional overlays
in nix to ensure that nix flake overrides, e.g. `nix build
--override-input kevm/k-framework v1.0.0`, do not only override the nix
flake but also the python package that was locked in `poetry.lock`.

With the migration to `uv2nix`, this overlay was removed, causing
confusion and overhead when wanting to override python code using nix
flake overrides. This PR re-introduces these overlays that can be used
in dependent nix flakes to re-establish this behaviour. Due to `uv2nix`
using the newer nix python packaging tooling `pyproject-nix`, this code
adds more complexity in comparison to `poetry2nix`, though it
nonetheless achieves the same result.

There will be similar PRs in other repositories introducing and using
these pyproject-nix overlays.