Streaming support for JsonRpcServer (#4879)
This PR extends the JsonRpcServer with support for streaming.
JsonRpcMethods can now return a stream of bytes (Iterator[bytes])
instead of a JSON-serializable value.
The method must ensure that the produced stream is already JSON-encoded,
as no extra validation is performed.
This allows us to transfer large responses without buffering the entire
response in memory.

This is needed by the kontrol-node, when transferring large execution
traces.